requirements.txt
django==1.11
djangorestframework==3.9.4
boto3==1.12.36